requirements.txt
streamlit
pyomo
groq
python-dotenv
fipy
matplotlib
fluidsim[fft,mpi]>=0.8.4
matplotlib>=3.5

